refactoring change - renamed "ethereum" folder to "truffle"
