+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/john/project/open-air-dapp/src/index.js":"1","/home/john/project/open-air-dapp/src/reportWebVitals.js":"2","/home/john/project/open-air-dapp/src/App.js":"3","/home/john/project/open-air-dapp/src/history.js":"4","/home/john/project/open-air-dapp/src/components/Home.js":"5","/home/john/project/open-air-dapp/src/components/NotFound.js":"6","/home/john/project/open-air-dapp/src/web3/web3.js":"7","/home/john/project/open-air-dapp/src/web3/openAirContract.js":"8","/home/john/project/open-air-dapp/src/components/OpenAirComponent.js":"9","/home/john/project/open-air-dapp/src/web3/opinionTokenContract.js":"10","/home/john/project/open-air-dapp/src/web3/speechContract.js":"11"},{"size":502,"mtime":1649388639797,"results":"12","hashOfConfig":"13"},{"size":362,"mtime":1649024763250,"results":"14","hashOfConfig":"13"},{"size":992,"mtime":1649117711584,"results":"15","hashOfConfig":"13"},{"size":87,"mtime":1649024763250,"results":"16","hashOfConfig":"13"},{"size":1401,"mtime":1649095662563,"results":"17","hashOfConfig":"13"},{"size":181,"mtime":1649024763250,"results":"18","hashOfConfig":"13"},{"size":168,"mtime":1649024763250,"results":"19","hashOfConfig":"13"},{"size":258,"mtime":1649211131801,"results":"20","hashOfConfig":"13"},{"size":18411,"mtime":1649430659075,"results":"21","hashOfConfig":"13"},{"size":287,"mtime":1649129424444,"results":"22","hashOfConfig":"13"},{"size":252,"mtime":1649216455861,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1i4q8nx",{"filePath":"26","messages":"27","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"31","messages":"32","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"33","messages":"34","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"35","messages":"36","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"37","messages":"38","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"39","messages":"40","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"41","messages":"42","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"45","messages":"46","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"/home/john/project/open-air-dapp/src/index.js",[],"/home/john/project/open-air-dapp/src/reportWebVitals.js",[],["47","48"],"/home/john/project/open-air-dapp/src/App.js",[],"/home/john/project/open-air-dapp/src/history.js",[],"/home/john/project/open-air-dapp/src/components/Home.js",[],"/home/john/project/open-air-dapp/src/components/NotFound.js",[],"/home/john/project/open-air-dapp/src/web3/web3.js",[],"/home/john/project/open-air-dapp/src/web3/openAirContract.js",[],"/home/john/project/open-air-dapp/src/components/OpenAirComponent.js",[],"/home/john/project/open-air-dapp/src/web3/opinionTokenContract.js",[],"/home/john/project/open-air-dapp/src/web3/speechContract.js",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},"no-native-reassign",["53"],"no-negated-in-lhs",["54"],"no-global-assign","no-unsafe-negation"]
+[{"/home/john/project/open-air-dapp/src/index.js":"1","/home/john/project/open-air-dapp/src/reportWebVitals.js":"2","/home/john/project/open-air-dapp/src/App.js":"3","/home/john/project/open-air-dapp/src/history.js":"4","/home/john/project/open-air-dapp/src/components/Home.js":"5","/home/john/project/open-air-dapp/src/components/NotFound.js":"6","/home/john/project/open-air-dapp/src/web3/web3.js":"7","/home/john/project/open-air-dapp/src/web3/openAirContract.js":"8","/home/john/project/open-air-dapp/src/components/OpenAirComponent.js":"9","/home/john/project/open-air-dapp/src/web3/opinionTokenContract.js":"10","/home/john/project/open-air-dapp/src/web3/speechContract.js":"11"},{"size":502,"mtime":1649388639797,"results":"12","hashOfConfig":"13"},{"size":362,"mtime":1649024763250,"results":"14","hashOfConfig":"13"},{"size":992,"mtime":1649117711584,"results":"15","hashOfConfig":"13"},{"size":87,"mtime":1649024763250,"results":"16","hashOfConfig":"13"},{"size":1400,"mtime":1649599719326,"results":"17","hashOfConfig":"13"},{"size":181,"mtime":1649024763250,"results":"18","hashOfConfig":"13"},{"size":168,"mtime":1649024763250,"results":"19","hashOfConfig":"13"},{"size":257,"mtime":1649599628876,"results":"20","hashOfConfig":"13"},{"size":18411,"mtime":1649430659075,"results":"21","hashOfConfig":"13"},{"size":286,"mtime":1649599602496,"results":"22","hashOfConfig":"13"},{"size":251,"mtime":1649599644266,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1i4q8nx",{"filePath":"26","messages":"27","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"31","messages":"32","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"33","messages":"34","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"37","messages":"38","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"39","messages":"40","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/john/project/open-air-dapp/src/index.js",[],"/home/john/project/open-air-dapp/src/reportWebVitals.js",[],["47","48"],"/home/john/project/open-air-dapp/src/App.js",[],"/home/john/project/open-air-dapp/src/history.js",[],"/home/john/project/open-air-dapp/src/components/Home.js",[],"/home/john/project/open-air-dapp/src/components/NotFound.js",[],"/home/john/project/open-air-dapp/src/web3/web3.js",[],"/home/john/project/open-air-dapp/src/web3/openAirContract.js",[],"/home/john/project/open-air-dapp/src/components/OpenAirComponent.js",[],"/home/john/project/open-air-dapp/src/web3/opinionTokenContract.js",[],"/home/john/project/open-air-dapp/src/web3/speechContract.js",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},"no-native-reassign",["53"],"no-negated-in-lhs",["54"],"no-global-assign","no-unsafe-negation"]
