+++ src/truffle/contracts/SharedStructs.sol
