+++ src/web3/opinionTokenContract.js
@@ -1,5 +1,5 @@
 import { web3 } from "./web3"
-import OpinionToken from '../ethereum/build/contracts/OpinionToken.json'
+import OpinionToken from '../truffle/build/contracts/OpinionToken.json'
 
 export function getOpinionTokenInstance(contractAddress) {
     let opinionTokenInstance = new web3.eth.Contract(OpinionToken.abi, contractAddress);
